CSHARP-4445: Support 64-bit values for Skip and Limit/Take.

+++ src/MongoDB.Driver/AggregateFluent.cs
@@ -147,7 +147,7 @@ namespace MongoDB.Driver
             return WithPipeline(_pipeline.Group(group));
         }
 
-        public override IAggregateFluent<TResult> Limit(int limit)
+        public override IAggregateFluent<TResult> Limit(long limit)
         {
             return WithPipeline(_pipeline.Limit(limit));
         }
@@ -278,7 +278,7 @@ namespace MongoDB.Driver
             return WithPipeline(_pipeline.SetWindowFields(partitionBy, sortBy, output));
         }
 
-        public override IAggregateFluent<TResult> Skip(int skip)
+        public override IAggregateFluent<TResult> Skip(long skip)
         {
             return WithPipeline(_pipeline.Skip(skip));
         }
